fix(desktop): separate sketchpad runtime sources and frame state
Generated-By: PostHog Desktop
Task-Id: 617b374c-f922-473b-96a7-f61b557c16ea
