+++ products/desktop/apps/code/src/shared/vite-raw-modules.d.ts
@@ -5,3 +5,13 @@ declare module "*.html?raw" {
   const content: string;
   export default content;
 }
+
+declare module "*.js?raw" {
+  const content: string;
+  export default content;
+}
+
+declare module "*.css?raw" {
+  const content: string;
+  export default content;
+}
