+++ products/desktop/packages/core/src/sketchpad/frame/rawFrameSources.d.ts
@@ -1,0 +1,8 @@
+declare module "*.js?raw" {
+  const source: string;
+  export default source;
+}
+declare module "*.css?raw" {
+  const source: string;
+  export default source;
+}
