+++ products/desktop/packages/ui/src/features/sketchpad/library/templates/trend-chart.tsx.txt
@@ -1,12 +1,4 @@
 import {
-  chartStates,
-  chartTooltip,
-  eventField,
-  labelField,
-  selectClass,
-} from "./chartParts";
-
-export const trendChartCode = `import {
   formatCompact,
   hogqlString,
   useDateRange,
@@ -41,7 +33,7 @@ const DEFAULTS = {
 };
 
 const SELECT_CLASS =
-  "${selectClass}";
+  "w-full rounded-(--radius-2) border border-border bg-card px-2 py-1 text-[12px] text-foreground outline-none";
 
 const MEASURES = [
   { value: "events", label: "Events" },
@@ -76,7 +68,21 @@ function dayLabel(day) {
   return (Number(parts[2]) + " " + (MONTHS[Number(parts[1]) - 1] ?? "")).trim();
 }
 
-${chartTooltip("dayLabel(label)")}
+function ChartTooltip({ active, payload, label, unit }) {
+  if (!active || !payload || payload.length === 0) {
+    return null;
+  }
+  return (
+    <div className="rounded-(--radius-2) border border-border bg-popover px-2.5 py-1.5 shadow-md">
+      <p className="max-w-56 truncate text-[11px] text-muted-foreground">
+        {dayLabel(label)}
+      </p>
+      <p className="font-medium text-[13px] tabular-nums">
+        {Number(payload[0].value).toLocaleString()} {unit}
+      </p>
+    </div>
+  );
+}
 
 export default function TrendChart({ fragmentId }) {
   const [settings, setSettings] = useFragmentSettings(fragmentId, DEFAULTS);
@@ -134,7 +140,24 @@ export default function TrendChart({ fragmentId }) {
 
       {open ? (
         <div className="flex min-h-0 flex-1 flex-col gap-2 overflow-auto">
-          ${eventField}
+          <Field>
+  <FieldLabel>Event</FieldLabel>
+  <select
+    aria-label="Event"
+    value={settings.event}
+    onChange={(change) => setSettings({ event: change.target.value })}
+    className={SELECT_CLASS}
+  >
+    {events.names.indexOf(settings.event) === -1 ? (
+      <option value={settings.event}>{settings.event}</option>
+    ) : null}
+    {events.names.map((name) => (
+      <option key={name} value={name}>
+        {name}
+      </option>
+    ))}
+  </select>
+</Field>
           <div className="flex gap-2">
             <Field className="flex-1">
               <FieldLabel>Count</FieldLabel>
@@ -167,11 +190,33 @@ export default function TrendChart({ fragmentId }) {
               </select>
             </Field>
           </div>
-          ${labelField}
+          <Field>
+  <FieldLabel>Label</FieldLabel>
+  <Input
+    value={settings.label}
+    className="h-7 text-[13px]"
+    onChange={(change) => setSettings({ label: change.target.value })}
+  />
+</Field>
         </div>
       ) : (
         <div className="min-h-0 flex-1">
-          ${chartStates}
+          {result.loading ? <Skeleton className="h-full w-full" /> : null}
+{result.error ? (
+  <div className="flex flex-col items-start gap-2">
+    <p className="text-xs text-destructive">This chart did not load.</p>
+    <Button variant="outline" size="sm" onClick={result.retry}>
+      Try again
+    </Button>
+  </div>
+) : null}
+{empty ? (
+  <div className="flex h-full items-center justify-center">
+    <p className="text-[13px] text-muted-foreground">
+      No events in this date range.
+    </p>
+  </div>
+) : null}
           {!result.loading && !result.error && rows.length > 0 ? (
             <ResponsiveContainer height="100%" width="100%">
               <AreaChart
@@ -233,4 +278,3 @@ export default function TrendChart({ fragmentId }) {
     </div>
   );
 }
-`;
