+++ products/desktop/packages/ui/src/features/sketchpad/runtime/useSketchpadFrame.test.ts
@@ -2,8 +2,6 @@ import {
   emptySketchpadSnapshot,
   SKETCHPAD_CHANNEL,
   SKETCHPAD_FRAME_TO_HOST_CHANNEL,
-  SKETCHPAD_HOST_TO_FRAME_CHANNEL,
-  type SketchpadDataMethod,
   type SketchpadFrameToHostMessage,
   type SketchpadSnapshot,
 } from "@posthog/shared";
@@ -45,34 +43,36 @@ function mountFrame(snapshot: SketchpadSnapshot, sketchpadId = "board") {
   const frame = document.createElement("webview") as SketchpadWebviewElement;
   frame.send = send;
   const queryClient = new QueryClient();
-  const { result, unmount } = renderHook(() =>
-    useSketchpadFrame({
-      sketchpadId,
-      frameElement: frame,
-      theme: "light",
-      queryClient,
-      getSnapshot: () => snapshot,
-      applyLocal: vi.fn(),
-      reportCaret: vi.fn(),
-      events: {
-        onExitFocus: vi.fn(),
-        onReady: vi.fn(),
-        onFragmentRendered: vi.fn(),
-        onFragmentError: vi.fn(),
-        onStateChanged: vi.fn(),
-        onWheel: vi.fn(),
-        onBackgroundPointer: vi.fn(),
-        onFragmentPointerDown: vi.fn(),
-        onPointerMove: vi.fn(),
-        onPointerLeave: vi.fn(),
-      },
-    }),
+  const { result, unmount, rerender } = renderHook(
+    ({ element }) =>
+      useSketchpadFrame({
+        sketchpadId,
+        frameElement: element,
+        theme: "light",
+        queryClient,
+        getSnapshot: () => snapshot,
+        applyLocal: vi.fn(),
+        reportCaret: vi.fn(),
+        events: {
+          onExitFocus: vi.fn(),
+          onReady: vi.fn(),
+          onFragmentRendered: vi.fn(),
+          onFragmentError: vi.fn(),
+          onStateChanged: vi.fn(),
+          onWheel: vi.fn(),
+          onBackgroundPointer: vi.fn(),
+          onFragmentPointerDown: vi.fn(),
+          onPointerMove: vi.fn(),
+          onPointerLeave: vi.fn(),
+        },
+      }),
+    { initialProps: { element: frame } },
   );
   act(() =>
     sendFromFrame(frame, { channel: SKETCHPAD_CHANNEL, type: "ready" }),
   );
   send.mockClear();
-  return { frame, send, result, unmount, queryClient };
+  return { frame, send, result, unmount, queryClient, rerender };
 }
 
 it("limits compilation requests and cancels them when the frame closes", async () => {
@@ -106,55 +106,6 @@ it("limits compilation requests and cancels them when the frame closes", async (
   queryClient.clear();
 });
 
-it.each([false, true])(
-  "bounds queued reads and handles a closed frame: %s",
-  async (close) => {
-    vi.useFakeTimers();
-    const { frame, send, unmount, queryClient } = mountFrame(
-      emptySketchpadSnapshot(),
-      `read-queue-${close}`,
-    );
-    let active = 0;
-    let peak = 0;
-    query.mockReset().mockImplementation(() => {
-      peak = Math.max(peak, ++active);
-      return new Promise((resolve) =>
-        setTimeout(() => {
-          active--;
-          resolve({ results: [] });
-        }, 20_000),
-      );
-    });
-    try {
-      await act(async () => {
-        for (let index = 0; index < 16; index++) {
-          sendFromFrame(frame, {
-            channel: SKETCHPAD_CHANNEL,
-            type: "data-request",
-            id: String(index),
-            method: "query",
-            payload: { hogql: `select ${index}` },
-          });
-        }
-      });
-      expect(query).toHaveBeenCalledTimes(8);
-      if (close) unmount();
-      await act(() => vi.advanceTimersByTimeAsync(40_000));
-      expect(peak).toBe(8);
-      expect(query).toHaveBeenCalledTimes(close ? 8 : 16);
-      const replies = send.mock.calls
-        .map(([, message]) => message)
-        .filter((message) => message.type === "data-response");
-      expect(replies).toHaveLength(close ? 0 : 16);
-      expect(replies.every((message) => message.ok)).toBe(true);
-    } finally {
-      unmount();
-      queryClient.clear();
-      vi.useRealTimers();
-    }
-  },
-);
-
 it("sends changed data without repeating unchanged source or state", () => {
   const toJSON = vi.fn(() => ({ value: "unchanged" }));
   const fragment = {
@@ -166,6 +117,8 @@ it("sends changed data without repeating unchanged source or state", () => {
     z: 0,
     code: "export default () => null",
     codeVersion: 1,
+    surface: "card" as const,
+    hidden: false,
   };
   const previous = {
     schemaVersion: 1 as const,
@@ -235,35 +188,36 @@ it("sends changed data without repeating unchanged source or state", () => {
   ]);
 });
 
-it.each<[SketchpadDataMethod, number, boolean]>([
-  ["stateEditText", 20_000, true],
-  ["stateEditText", 60_000, false],
-  ["query", 20_000, false],
-])("bounds %s requests with %i entry IDs", async (method, count, ok) => {
-  const { frame, send } = mountFrame(emptySketchpadSnapshot());
-  await act(async () => {
-    sendFromFrame(frame, {
-      channel: SKETCHPAD_CHANNEL,
-      type: "data-request",
-      id: "request",
-      method,
-      payload: {
-        key: "note",
-        base: "a".repeat(count),
-        next: "b".repeat(count),
-        baseIds: Array.from(
-          { length: count },
-          (_, index) => `${"a".repeat(32)}-${index}`,
-        ),
-      },
-    });
-  });
-  expect(send).toHaveBeenCalledWith(
-    SKETCHPAD_HOST_TO_FRAME_CHANNEL,
-    expect.objectContaining({
-      type: "data-response",
-      id: "request",
-      ok,
-    }),
+it("resets readiness and caret delivery when the frame element changes", () => {
+  const { result, rerender, unmount, queryClient } = mountFrame(
+    emptySketchpadSnapshot(),
   );
+  const carets = [
+    {
+      clientId: "person",
+      key: "note",
+      anchor: null,
+      focus: null,
+      name: "Person",
+      color: "#000000",
+      textColor: "#ffffff",
+    },
+  ];
+  result.current.setCarets(carets);
+  const frame = document.createElement("webview") as SketchpadWebviewElement;
+  frame.send = vi.fn();
+  rerender({ element: frame });
+  expect(result.current.ready).toBe(false);
+  result.current.setCarets(carets);
+  expect(frame.send).not.toHaveBeenCalled();
+  act(() =>
+    sendFromFrame(frame, { channel: SKETCHPAD_CHANNEL, type: "ready" }),
+  );
+  result.current.setCarets(carets);
+  expect(frame.send).toHaveBeenCalledWith(
+    "posthog-sketchpad-host",
+    expect.objectContaining({ type: "set-carets", carets }),
+  );
+  unmount();
+  queryClient.clear();
 });
