+++ products/desktop/packages/ui/src/features/sketchpad/runtime/useSketchpadFrame.ts
@@ -1,17 +1,8 @@
-import { resolveService } from "@posthog/di/container";
+import { buildSketchpadFrameDocument } from "@posthog/core/sketchpad/frameDocument";
 import {
-  HOST_TRPC_CLIENT,
-  type HostTrpcClient,
-} from "@posthog/host-router/client";
-import {
-  estimateJsonBytes,
   fragmentsEqual,
   type HostToSketchpadFrameMessage,
-  isField,
-  isReservedStateKey,
-  isSafePostHogUrl,
   SKETCHPAD_CHANNEL,
-  SKETCHPAD_STATE_KEY_MAX_CHARS,
   type SketchpadFrameCaret,
   type SketchpadFrameToHostMessage,
   type SketchpadOp,
@@ -22,10 +13,7 @@ import {
   sketchpadFrameToHostMessageSchema,
 } from "@posthog/shared";
 import { useSketchpadApi } from "@posthog/ui/features/sketchpad/hooks/useSketchpadApi";
-import { spendSketchpadWrite } from "@posthog/ui/features/sketchpad/runtime/sketchpadDataBridge";
 import { fieldMessageValue } from "@posthog/ui/features/sketchpad/runtime/sketchpadFieldMessages";
-import { SKETCHPAD_TOO_MANY_READS_AT_ONCE } from "@posthog/ui/features/sketchpad/sketchpadCopy";
-import { logger } from "@posthog/ui/shell/logger";
 import { openExternalUrl } from "@posthog/ui/shell/openExternal";
 import { useHostCapabilities } from "@posthog/ui/shell/useHostCapabilities";
 import type { QueryClient } from "@tanstack/react-query";
@@ -37,25 +25,13 @@ import {
   useRef,
   useState,
 } from "react";
-import {
-  handleSketchpadDataRequest,
-  type SketchpadDataBridgeContext,
-} from "./sketchpadDataBridge";
-import {
-  buildSketchpadFrameDocument,
-  sketchpadFramePolicy,
-} from "./sketchpadFrameDocument";
+import { createSketchpadBudget } from "./sketchpadDataBridge";
 import {
   listenToSketchpadFrame,
   type SketchpadFrameElement,
   sendToSketchpadFrame,
 } from "./sketchpadFrameElement";
-
-const log = logger.scope("sketchpad-frame");
-
-const MAX_PENDING_DATA_REQUESTS = 200;
-const MAX_DATA_REQUEST_BYTES = 64 * 1024;
-const EXTERNAL_OPEN_MIN_INTERVAL_MS = 1_000;
+import { createSketchpadHostMessageRouter } from "./sketchpadHostMessageRouter";
 
 export interface SketchpadFrameEvents {
   onExitFocus(): void;
@@ -87,7 +63,6 @@ export interface UseSketchpadFrameOptions {
 
 export interface SketchpadFrameHandle {
   srcDoc: string;
-  documentReady: boolean;
   ready: boolean;
   post: (message: HostToSketchpadFrameMessage) => void;
   sendInit: (viewport: SketchpadViewport) => void;
@@ -112,26 +87,6 @@ export function useSketchpadFrame(
     () => buildSketchpadFrameDocument({ vendoredModules }),
     [vendoredModules],
   );
-  const [documentReady, setDocumentReady] = useState(!vendoredModules);
-
-  useEffect(() => {
-    if (!vendoredModules) return;
-    let cancelled = false;
-    void resolveService<HostTrpcClient>(HOST_TRPC_CLIENT)
-      .sketchpadFrame.registerDocument.mutate({
-        html: srcDoc,
-        csp: sketchpadFramePolicy(true),
-      })
-      .then(() => {
-        if (!cancelled) setDocumentReady(true);
-      })
-      .catch((error: unknown) => {
-        log.error("Could not register the board frame document", { error });
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, [srcDoc, vendoredModules]);
   const [ready, setReady] = useState(false);
   const readyRef = useRef(false);
 
@@ -223,6 +178,7 @@ export function useSketchpadFrame(
   const sentCarets = useRef("");
   const setCarets = useCallback(
     (carets: SketchpadFrameCaret[]): void => {
+      if (!readyRef.current) return;
       const json = stableJson(carets);
       if (json === sentCarets.current) return;
       sentCarets.current = json;
@@ -260,186 +216,27 @@ export function useSketchpadFrame(
   );
 
   useLayoutEffect(() => {
-    let activeRequests = 0;
-    let lastExternalOpen = 0;
+    readyRef.current = false;
+    setReady(false);
+    fromFrame.current.clear();
+    sentCarets.current = "";
+    const budget = createSketchpadBudget(options.sketchpadId);
     const controller = new AbortController();
-
-    const reply = (
-      id: string,
-      ok: boolean,
-      result?: unknown,
-      error?: string,
-    ): void => {
-      if (controller.signal.aborted) return;
-      postRaw({
-        channel: SKETCHPAD_CHANNEL,
-        type: "data-response",
-        id,
-        ok,
-        result,
-        error,
-      });
-    };
-
-    const runDataRequest = async (
-      message: Extract<SketchpadFrameToHostMessage, { type: "data-request" }>,
-    ): Promise<void> => {
-      if (
-        activeRequests >= MAX_PENDING_DATA_REQUESTS ||
-        estimateJsonBytes(message.payload) >
-          (message.method === "stateEditText"
-            ? 2 * 1024 * 1024
-            : MAX_DATA_REQUEST_BYTES)
-      ) {
-        reply(message.id, false, undefined, SKETCHPAD_TOO_MANY_READS_AT_ONCE);
-        return;
-      }
-      activeRequests += 1;
-      const { sketchpadId, queryClient, getSnapshot, applyLocal, reportCaret } =
-        latest.current;
-      const ctx: SketchpadDataBridgeContext = {
-        sketchpadId,
-        queryClient,
-        getSnapshot,
-        applyLocal,
-        reportCaret,
-        signal: controller.signal,
-      };
-      try {
-        const result = await handleSketchpadDataRequest(
-          message.method,
-          message.payload,
-          ctx,
-        );
-        reply(message.id, true, result);
-      } catch (error) {
-        reply(
-          message.id,
-          false,
-          undefined,
-          error instanceof Error ? error.message : String(error),
-        );
-      } finally {
-        activeRequests -= 1;
-      }
-    };
-
-    let compiling = false;
-    const runCompileRequest = async (
-      message: Extract<
-        SketchpadFrameToHostMessage,
-        { type: "compile-request" }
-      >,
-    ): Promise<void> => {
-      if (compiling) {
-        reply(
-          message.id,
-          false,
-          undefined,
-          "A fragment compilation request is already active.",
-        );
-        return;
-      }
-      compiling = true;
-      try {
-        reply(
-          message.id,
-          true,
-          await api.compiled(
-            options.sketchpadId,
-            message.refs,
-            controller.signal,
-          ),
-        );
-      } catch (error) {
-        reply(
-          message.id,
-          false,
-          undefined,
-          error instanceof Error ? error.message : String(error),
-        );
-      } finally {
-        compiling = false;
-      }
-    };
-
-    const route = (message: SketchpadFrameToHostMessage): void => {
-      const events = latest.current.events;
-      switch (message.type) {
-        case "compile-request":
-          void runCompileRequest(message);
-          break;
-        case "ready":
-          readyRef.current = true;
-          setReady(true);
-          events.onReady();
-          break;
-        case "exit-focus":
-          events.onExitFocus();
-          break;
-        case "fragment-rendered":
-          events.onFragmentRendered(message.id);
-          break;
-        case "fragment-error":
-          events.onFragmentError(message.id, message.message, message.stack);
-          break;
-        case "state-changed":
-          if (
-            isReservedStateKey(message.key) ||
-            message.key.length > SKETCHPAD_STATE_KEY_MAX_CHARS
-          ) {
-            log.warn("Refused a fragment state key");
-            break;
-          }
-          if (!spendSketchpadWrite(options.sketchpadId)) {
-            log.warn("Paused a fragment that writes shared state too fast");
-            break;
-          }
-          if (isField(latest.current.getSnapshot().state[message.key])) break;
-          fromFrame.current.set(message.key, stableJson(message.value ?? null));
-          events.onStateChanged(message.key, message.value);
-          break;
-        case "policy-violation":
-          log.warn("A fragment tried a channel the board closes", {
-            directive: message.directive,
-            blocked: message.blocked,
-          });
-          break;
-        case "data-request":
-          void runDataRequest(message);
-          break;
-        case "wheel":
-          events.onWheel(message);
-          break;
-        case "background-pointer":
-          events.onBackgroundPointer(message);
-          break;
-        case "fragment-pointer-down":
-          events.onFragmentPointerDown(message);
-          break;
-        case "pointer-move":
-          events.onPointerMove(message.clientX, message.clientY);
-          break;
-        case "pointer-leave":
-          events.onPointerLeave();
-          break;
-        case "open-external": {
-          const now = Date.now();
-          if (!isSafePostHogUrl(message.url)) {
-            log.warn("Blocked non-PostHog fragment external URL");
-          } else if (navigator.userActivation?.isActive !== true) {
-            log.warn("Ignored fragment external URL open without interaction");
-          } else if (now - lastExternalOpen < EXTERNAL_OPEN_MIN_INTERVAL_MS) {
-            log.warn("Throttled fragment external URL open");
-          } else {
-            lastExternalOpen = now;
-            openExternalUrl(message.url);
-          }
-          break;
-        }
-      }
-    };
-
+    const route = createSketchpadHostMessageRouter({
+      post: postRaw,
+      signal: controller.signal,
+      callbacks: () => latest.current,
+      budget,
+      compiled: api.compiled.bind(api),
+      hasUserActivation: () => navigator.userActivation?.isActive === true,
+      openExternal: openExternalUrl,
+      onReady: () => {
+        readyRef.current = true;
+        setReady(true);
+      },
+      onStateEcho: (key, value) =>
+        fromFrame.current.set(key, stableJson(value)),
+    });
     if (!frameElement) return;
     const stopListening = listenToSketchpadFrame(frameElement, (data) => {
       const parsed = sketchpadFrameToHostMessageSchema.safeParse(data);
@@ -450,7 +247,7 @@ export function useSketchpadFrame(
       controller.abort();
       stopListening();
     };
-  }, [frameElement, postRaw, options.sketchpadId, api]);
+  }, [frameElement, postRaw, api, options.sketchpadId]);
 
   useEffect(() => {
     post({
@@ -462,7 +259,6 @@ export function useSketchpadFrame(
 
   return {
     srcDoc,
-    documentReady,
     ready,
     post,
     sendInit,
